1mM leucine processing, with some processing changes
The big changes:

1. Programs 0 and 1 now calculate variance across the image to create a
binary mask that rejects any stationary cells. This allows the point
picker to be more greedy. I still think the particles picked are not the
center of rotation, so a local maximum detector of the means should be
employed. Will do that next.

2. Program 1 uses a new function called process_kymograph2 which filters
the kymograph by subtracting the mean row value from each row. This
removes streaks that go across the entire stream due to interference
from other cells or mismatch of the tether point.
3. Program 1 saves two .tif files: before and after processing. I've
written a MATLAB script that pulls out kymograph direction quite
reliably, so I will be working that into 2_kymographs_to_traces next. :)

I have to say that working with this code has been pretty easy so far,
so thanks for commenting well!